+++ src/folding/nova/decider_eth_circuit.rs
@@ -21,8 +21,7 @@ use crate::ccs::r1cs::R1CS;
 use crate::commitment::{pedersen::Params as PedersenParams, CommitmentProver};
 use crate::folding::nova::{
     circuits::{CommittedInstanceVar, CF1, CF2},
-    ivc::IVC,
-    CommittedInstance, Witness,
+    CommittedInstance, Nova, Witness,
 };
 use crate::frontend::FCircuit;
 use crate::utils::gadgets::{
@@ -179,7 +178,7 @@ where
 
 /// Circuit that implements the in-circuit checks needed for the onchain (Ethereum's EVM)
 /// verification.
-pub struct DeciderCircuit<C1, GC1, C2, GC2, CP1, CP2>
+pub struct DeciderEthCircuit<C1, GC1, C2, GC2, CP1, CP2>
 where
     C1: CurveGroup,
     GC1: CurveVar<C1, CF2<C1>>,
@@ -220,7 +219,7 @@ where
     pub cf_U_i: Option<CommittedInstance<C2>>,
     pub cf_W_i: Option<Witness<C2>>,
 }
-impl<C1, GC1, C2, GC2, CP1, CP2> DeciderCircuit<C1, GC1, C2, GC2, CP1, CP2>
+impl<C1, GC1, C2, GC2, CP1, CP2> DeciderEthCircuit<C1, GC1, C2, GC2, CP1, CP2>
 where
     C1: CurveGroup,
     C2: CurveGroup,
@@ -230,8 +229,8 @@ where
     // enforce that the CP2 is Pedersen commitment, since we're at Ethereum's EVM decider
     CP2: CommitmentProver<C2, Params = PedersenParams<C2>>,
 {
-    pub fn from_ivc<FC: FCircuit<C1::ScalarField>>(
-        ivc: IVC<C1, GC1, C2, GC2, FC, CP1, CP2>,
+    pub fn from_nova<FC: FCircuit<C1::ScalarField>>(
+        nova: Nova<C1, GC1, C2, GC2, FC, CP1, CP2>,
     ) -> Self {
         Self {
             _c1: PhantomData,
@@ -241,27 +240,27 @@ where
             _cp1: PhantomData,
             _cp2: PhantomData,
 
-            E_len: ivc.W_i.E.len(),
-            cf_E_len: ivc.cf_W_i.E.len(),
-            r1cs: ivc.r1cs,
-            cf_r1cs: ivc.cf_r1cs,
-            cf_pedersen_params: ivc.cf_cm_params,
-            poseidon_config: ivc.poseidon_config,
-            i: Some(ivc.i),
-            z_0: Some(ivc.z_0),
-            z_i: Some(ivc.z_i),
-            u_i: Some(ivc.u_i),
-            w_i: Some(ivc.w_i),
-            U_i: Some(ivc.U_i),
-            W_i: Some(ivc.W_i),
-            cf_U_i: Some(ivc.cf_U_i),
-            cf_W_i: Some(ivc.cf_W_i),
+            E_len: nova.W_i.E.len(),
+            cf_E_len: nova.cf_W_i.E.len(),
+            r1cs: nova.r1cs,
+            cf_r1cs: nova.cf_r1cs,
+            cf_pedersen_params: nova.cf_cm_params,
+            poseidon_config: nova.poseidon_config,
+            i: Some(nova.i),
+            z_0: Some(nova.z_0),
+            z_i: Some(nova.z_i),
+            u_i: Some(nova.u_i),
+            w_i: Some(nova.w_i),
+            U_i: Some(nova.U_i),
+            W_i: Some(nova.W_i),
+            cf_U_i: Some(nova.cf_U_i),
+            cf_W_i: Some(nova.cf_W_i),
         }
     }
 }
 
 impl<C1, GC1, C2, GC2, CP1, CP2> ConstraintSynthesizer<CF1<C1>>
-    for DeciderCircuit<C1, GC1, C2, GC2, CP1, CP2>
+    for DeciderEthCircuit<C1, GC1, C2, GC2, CP1, CP2>
 where
     C1: CurveGroup,
     C2: CurveGroup,
@@ -452,9 +451,11 @@ pub mod tests {
     use ark_vesta::{constraints::GVar as GVar2, Projective as Projective2};
 
     use crate::commitment::pedersen::Pedersen;
-    use crate::folding::nova::ivc::tests::get_pedersen_params_len;
+    use crate::folding::nova::tests::get_test_pedersen_params;
+    use crate::folding::nova::{ProverParams, VerifierParams};
     use crate::frontend::tests::{CubicFCircuit, CustomFCircuit, WrapperCircuit};
     use crate::transcript::poseidon::tests::poseidon_test_config;
+    use crate::FoldingScheme;
 
     use crate::ccs::r1cs::{extract_r1cs, extract_w_x};
     use crate::ccs::r1cs::{
@@ -619,14 +620,22 @@ pub mod tests {
         let F_circuit = CubicFCircuit::<Fr>::new(());
         let z_0 = vec![Fr::from(3_u32)];
 
-        let (pedersen_len, cf_pedersen_len) =
-            get_pedersen_params_len::<CubicFCircuit<Fr>>(&poseidon_config, F_circuit).unwrap();
-        // generate the Pedersen params
-        let pedersen_params = Pedersen::<Projective>::new_params(&mut rng, pedersen_len);
-        let cf_pedersen_params = Pedersen::<Projective2>::new_params(&mut rng, cf_pedersen_len);
+        // get the CM & CF_CM len
+        let (pedersen_params, cf_pedersen_params) =
+            get_test_pedersen_params::<_, Projective, GVar, Projective2, GVar2, CubicFCircuit<Fr>>(
+                &mut rng,
+                &poseidon_config,
+                F_circuit,
+            );
 
-        // generate an IVC and do a step of it
-        let mut ivc = IVC::<
+        let prover_params =
+            ProverParams::<Projective, Projective2, Pedersen<Projective>, Pedersen<Projective2>> {
+                poseidon_config: poseidon_config.clone(),
+                cm_params: pedersen_params,
+                cf_cm_params: cf_pedersen_params,
+            };
+
+        type NOVA = Nova<
             Projective,
             GVar,
             Projective2,
@@ -634,26 +643,37 @@ pub mod tests {
             CubicFCircuit<Fr>,
             Pedersen<Projective>,
             Pedersen<Projective2>,
-        >::new(
-            poseidon_config,
-            pedersen_params,
-            cf_pedersen_params,
-            F_circuit,
-            z_0.clone(),
+        >;
+
+        // generate a Nova instance and do a step of it
+        let mut nova = NOVA::init(&prover_params, F_circuit, z_0.clone()).unwrap();
+        nova.prove_step().unwrap();
+        let ivc_v = nova.clone();
+        let verifier_params = VerifierParams::<Projective, Projective2> {
+            poseidon_config: poseidon_config.clone(),
+            r1cs: ivc_v.r1cs,
+            cf_r1cs: ivc_v.cf_r1cs,
+        };
+        NOVA::verify(
+            verifier_params,
+            z_0,
+            ivc_v.z_i,
+            Fr::one(),
+            (ivc_v.U_i, ivc_v.W_i),
+            (ivc_v.u_i, ivc_v.w_i),
+            (ivc_v.cf_U_i, ivc_v.cf_W_i),
         )
         .unwrap();
-        ivc.prove_step().unwrap();
-        ivc.verify(z_0, 1).unwrap();
 
-        // load the DeciderCircuit from the generated IVC
-        let decider_circuit = DeciderCircuit::<
+        // load the DeciderEthCircuit from the generated Nova instance
+        let decider_circuit = DeciderEthCircuit::<
             Projective,
             GVar,
             Projective2,
             GVar2,
             Pedersen<Projective>,
             Pedersen<Projective2>,
-        >::from_ivc(ivc);
+        >::from_nova(nova);
 
         let cs = ConstraintSystem::<Fr>::new_ref();
